Convert to function component
