+++ tsconfig.json
@@ -29,7 +29,8 @@
   "include": [
     "src/**/*.ts",
     "bench/**/*.ts",
-    "bench/**/*.js"
+    "bench/**/*.js",
+    "bench/**/*.tsx"
   ],
   "exclude": [
     "node_modules",
